+++ CHANGES.rst
@@ -5,6 +5,10 @@ Changelog
 ------------------
 
 - #2652 Fix Subgroups sort by Sort Key
+- #2663 CSS for some add-on based forms
+- #2661 Allow to recalculate an analysis result
+- #2659 Refactor sample title + decsription into viewlets
+- #2660 Add logging if external calculation module import failed
 - #2654 Show Batch title or ID in Sample reference field
 - #2657 Methods from analyses are not updated on instrument change in worksheet
 - #2656 Fix AnalysisProfile keyword validator fail with non-ascii value